fix: use `RUN::config` for run number access in `monitorRead.groovy`
