+++ qa-physics/Tools.groovy
@@ -1,5 +1,6 @@
 import org.jlab.clas.physics.LorentzVector
 import org.jlab.clas.physics.Vector3
+import org.jlab.io.hipo.HipoDataSource
 import groovy.json.JsonOutput
 
 class Tools {
@@ -189,4 +190,27 @@ class Tools {
   // print text output
   def printStatus = { str -> println "STATUS: $str" }
   def printError = { str -> System.err << "ERROR: $str\n" }
+
+
+  ////////////////
+  // HIPO FILES //
+  ////////////////
+  // get a run number from RUN::config
+  def getRunNumber(infile) {
+    def reader = new HipoDataSource()
+    def runnum = 0
+    def event
+    reader.open(infile)
+    while(reader.hasEvent()) {
+      event = reader.getNextEvent()
+      if(event.hasBank("RUN::config")) {
+        runnum = BigInteger.valueOf(event.getBank("RUN::config").getInt('run',0))
+        break
+      }
+    }
+    reader.close()
+    if(runnum<=0)
+      System.err.println("[ERROR]: run number not found for file $infile")
+    return runnum
+  }
 }
